Working Motor Driver Code
Working motor driver code, still needs to take signed height as input to be interpreted into direction(base on the sign) and height. Speed could be taken as input too or just choose a constant speed or make the speed vary while lowering/pulling (e.g.: slower in the beginning and towards reaching the wanted height).